Switched IDs to longs from ints

+++ java/src/com/blairtrump/blush/Instance.java
@@ -5,14 +5,14 @@ import java.util.Map;
 
 public class Instance {
 	String name;
-	Map<Integer, String> clientReplyQueues;
+	Map<Long, String> clientReplyQueues;
 	
 	public Instance(String name) {
 		this.name = name;
-		clientReplyQueues = new HashMap<Integer, String>();
+		clientReplyQueues = new HashMap<Long, String>();
 	}
 	
-	public boolean connect(Integer clientId, String replyQueue) {
+	public boolean connect(Long clientId, String replyQueue) {
 		boolean success = true;
 		try {
 			clientReplyQueues.put(clientId, replyQueue);
